Update inner-join.md (typo on inner join example)
I just want to suggest an edit I wasnt sure if it's a typo or not.

For the inner join example, it's says first two rows of the inner join, but it is showing three rows. I assumed it's a typo and  I just thought it was confusing at first and wanted to ask about it.